refactor, rename, calibrationtrigger on prefab
