+++ Assets/Scenes/GazeVisualizerScene.unity
@@ -153,13 +153,18 @@ PrefabInstance:
     - target: {fileID: 1409125293413158502, guid: 220628ae61044264f99c8bc2419ebd99,
         type: 3}
       propertyPath: DataClients.Array.size
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 1409125293413158502, guid: 220628ae61044264f99c8bc2419ebd99,
         type: 3}
       propertyPath: DataClients.Array.data[0]
       value: 
       objectReference: {fileID: 1992163580}
+    - target: {fileID: 1409125293413158502, guid: 220628ae61044264f99c8bc2419ebd99,
+        type: 3}
+      propertyPath: DataClients.Array.data[1]
+      value: 
+      objectReference: {fileID: 1443916687}
     - target: {fileID: 4625099936160350869, guid: 220628ae61044264f99c8bc2419ebd99,
         type: 3}
       propertyPath: m_LocalPosition.x
@@ -387,6 +392,86 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &1443916686
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 4657624760085729584, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_Name
+      value: Eye-Vive-EyeDataConverterSRanipal
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -0.30041963
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0.6611201
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0.07547766
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6953486736792480888, guid: f0f4ba2774dcea046987b82709fdc7cd,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: f0f4ba2774dcea046987b82709fdc7cd, type: 3}
+--- !u!114 &1443916687 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 5311309715874392284, guid: f0f4ba2774dcea046987b82709fdc7cd,
+    type: 3}
+  m_PrefabInstance: {fileID: 1443916686}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 4f27061cfab66a14ea683188127403a4, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &1992163579
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -616,3 +701,4 @@ SceneRoots:
   - {fileID: 2052440886}
   - {fileID: 323908166}
   - {fileID: 1992163579}
+  - {fileID: 1443916686}
