+++ Assets/com.edia.eye.vive/Runtime/Scripts/EyeCalibrationTriggerViveSRanipal.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+using Edia;
+
+namespace Edia.Eye.Vive {
+
+	public class EyeCalibrationTriggerViveSRanipal : EyeCalibrationTriggerBase {
+
+		public override void OnEvEyeCalibrationRequested(eParam e) {
+			Debug.Log(name + " OnEvEyeCalibrationRequested");
+
+			// VIVE
+			ViveSR.anipal.Eye.SRanipal_Eye.LaunchEyeCalibration();
+		}
+	}
+}
