--- Assets/com.edia.eye.vive/Runtime/Scripts/EyeGazeVisualizerSRanipal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: aaec49f236ae0d54e94752bf608fef98
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: ed2e144a4a003054085892bf7f1a66be, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
